+++ cmd/detect/const.go
@@ -1,0 +1,5 @@
+package detect
+
+import "time"
+
+const timeout = 100 * time.Millisecond
